Move test source code to src\test\java
