+++ src/test/java/com/mendix/audittrail/tests/TestAuditTrigger.java
@@ -1,4 +1,4 @@
-package test_crm.tests;
+package com.mendix.audittrail.tests;
 
 import com.mendix.core.CoreException;
 import com.mendix.systemwideinterfaces.core.IMendixObject;
@@ -8,8 +8,8 @@ import org.junit.Test;
 import audittrail.proxies.MemberType;
 import audittrail.proxies.TypeOfLog;
 import test_crm.proxies.ContactPerson;
-import test_crm.tests.actual.ActualLog;
-import test_crm.tests.expected.ExpectedLog;
+import com.mendix.audittrail.tests.actual.ActualLog;
+import com.mendix.audittrail.tests.expected.ExpectedLog;
 
 import static test_crm.proxies.ContactPerson.MemberNames.Firstname;
 import static test_crm.proxies.ContactPerson.MemberNames.Surname;
