+++ src/test/java/com/mendix/audittrail/tests/expected/ExpectedLogLineAttribute.java
@@ -1,4 +1,4 @@
-package test_crm.tests.expected;
+package com.mendix.audittrail.tests.expected;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
@@ -7,7 +7,7 @@ import com.mendix.core.CoreException;
 
 import audittrail.proxies.LogLine;
 import audittrail.proxies.MemberType;
-import test_crm.tests.actual.ActualLogLine;
+import com.mendix.audittrail.tests.actual.ActualLogLine;
 
 public class ExpectedLogLineAttribute extends ExpectedLogLine {
 	private final String oldValue;
